requirements.txt
asgiref==3.8.1
asttokens==3.0.0
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
cssselect2==0.8.0
decorator==5.2.1
Django==5.2.3
django-filter==25.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
executing==2.2.0
fonttools==4.59.0
iniconfig==2.1.0
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Markdown==3.8.2
matplotlib-inline==0.1.7
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.2
pyphen==0.17.2
pytest==8.4.1
pytest-django==4.11.1
sentry-sdk==2.34.1
sqlparse==0.5.3
stack-data==0.6.3
tinycss2==1.4.0
tinyhtml5==2.0.0
traitlets==5.14.3
urllib3==2.5.0
wcwidth==0.2.13
weasyprint==66.0
webencodings==0.5.1
zopfli==0.2.3.post1
